Add files required for Debian packaging
This adds a debian/* folder containing the basic informtion to
build a debian package using dpkg-buildpackage.
It also adds a .reuse folder containing a dep5 license file that
specifies the license for the debian folder.
Bot the debian folder and the .reuse folder are also listed as
exporte-ignore in the .gitattributes file