Support for building on Debian 9 & Ubuntu 16.04
We still want to support building the package on Debian 9. This means
that we have to depend on an older version of debhelper and also that we
have to use the debian/compat file instead of depending on
debhelper-compat.
This commit also enables building of the package using github workflows.

Closes #5