MAINT: Restructure data files
The intelmq-api-config.json file was renamed to config.json and is now
directly installed to /etc/intelmq-api/config.json
The apache.conf file was renamed to apache-virtualhost.conf. It is also
installed in /etc/intelmq-api and the Debian package now creates a
symlink to this file in /etc/apache2/sites-available.
The setup.py now also includes a package_data dict whereby the
intelmq-api.wsgi is also installed in the package directory.