ENH: Add handling of default configuration files
IntelMQ-API crashed if the used configuration file did not exists. This
change fixes that by checking for the existence of the file and using
default values if it does not exist or is not a JSON file.
In addition to that IntelMQ-API now has a list of default paths for
configuration files, i.e. one does not need to set one by default.