ENH: Rewrite the configuration file handling
In IntelMQ 3 the defaults.conf and the pipeline.conf were dropped,
therefore access to those files via the API is not useful anymore.
At the same time, the runtime.conf now includes pipeline configuration
and is stored as a YAML file. Given those changes, the whole
configuration file handling had to be rewritten: There are now 3 API
endpoints:
* one to read (get) and write (post) the runtime configuration
* one to read (get) and write (post) the positions configuration
* one to read (get) the harmonization information

Everything is still done using JSON objects, the conversion to YAML is
done internally.
The removal of the save_file method also includes the removal of some
sanity checks that ran before the file was saved. Some of those should
probably be part of intelmq itself (i.e. checking for allowed characters
in bot names).