Dropped Python 3.6. Checked code style. New deps
As an intro for transition to the FastAPI and to be in
sync with the IntelMQ 3.1, the Python 3.6 was dropped.
In addition, the code style checks was introduced to the
CI. The config is the same as for IntelMQ core.

Dependencies for FastAPI was also added to check CI compatibility.