ENH: Package updated to setup new API
The FastAPI-based API is now an ASGI application, that isn't
natively supported by Apache. Instead, a recommended FastAPI
setup with web server Gunicorn is used. To keep backward compatibility,
Apache is still used as a Proxy. Apache and Gunicorn communicate
using unix socket.

The package installation creates intelmq-api service and socket
for systemd, and configure the Apache2 to communicate with them.
The INTELMQ_API_CONFIG env can now be set in the service unit file.

Provided configuration is an MVP and should be reviewed in the
production environement.